refactor: :recycle: refactor modals
