+++ ui/src/types/index.ts
@@ -7,3 +7,4 @@ export * from "./apiTypes"
 export * from "./sourceTypes"
 export * from "./destinationTypes"
 export * from "./formTypes"
+export * from "./modalTypes"
