+++ ui/src/types/modalTypes.ts
@@ -1,0 +1,3 @@
+export interface DeleteModalProps {
+	fromSource: boolean
+}
